ChemGCNs/requirements.txt
matplotlib==3.7.5
mendeleev==0.19.0
numpy==1.24.4
pandas==2.0.3
pyyaml==6.0.2
rdkit-pypi==2022.9.5
scikit-learn==1.3.2
rpy2==3.5.11
torchdata==0.7